feat: useRefresh
